Switch mainline to Python 3.x only
This has been a long time coming, and Marc Sommerhalder finally pestered
me enough to follow through. Python 3.x is now at the same point the
last version of Python 2.x was at: .7.

This commit doesn't change all that much that wasn't on the Python3
branch already: it updates the README, gets rid of some specical casing
and updates infrastructure.

Oh, and we're now at 1.0. Not that that you should ascribe much meaning
to that, but it's a big enough change to warrant a major bump.