Simplify re.match.
There is a simpler way to get the matched string from a re search.

However, since all the regexes we use are anchored with ^ and $, the
match can only ever be the entire string anyway, so we should really
just do

matches.append(wpart)

but maybe there's a reason for not doing this that I don't understand.